* Move a bunch of block related code into a separate file.
